+++ projects/Cyseal/Cyseal.vcxproj
@@ -115,6 +115,7 @@
     <ClInclude Include="src\render\buffer_visualization.h" />
     <ClInclude Include="src\render\gpu_culling.h" />
     <ClInclude Include="src\render\pathtracing\path_tracing_pass.h" />
+    <ClInclude Include="src\render\renderer_options.h" />
     <ClInclude Include="src\render\util\volatile_descriptor.h" />
     <ClInclude Include="src\rhi\buffer.h" />
     <ClInclude Include="src\rhi\descriptor_heap.h" />
